backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
feedparser==6.0.11
apscheduler==3.10.4
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
